requirements.txt
celery
Django>=1.10,<2.0
django-crispy-forms
django-debug-toolbar
django-cms

redis
requests
django-redis
six
django-anymail # Used for mandrill integration. See installation instruction at https://github.com/anymail/
psycopg2
django-hijack==2.1.4
django-hijack-admin==2.1.4
django-imagekit==4.0.2
django-ckeditor
Pillow==5.1.0
django-sortedm2m==1.5.0
django-autoslug==1.9.3
djangocms-picture
django-filer
django-nested-inline
django-nested-admin
django-admin-sortable
django-parler==1.9.2
django-grappelli
django-instagram
django-cart
django-constance[database]


django-smtp-ssl

boto3
django-storages
django-boto

solid_i18n
jsonfield